Move the class into the module
In addition to being more meaningful, this also helps with e.g. sphinx autodoc
